Don't allow delete the original locale emails, don't change seotitle on content page/faq title edit

+++ classes/controller/panel/content.php
@@ -173,8 +173,6 @@ class Controller_Panel_Content extends Auth_Controller {
                 }
                 // if status is not checked, it is not set as POST response
                 $content->status = (isset($p['status']))?1:0;
-                if ($type!='email')//email we do not update the seoname if not wont find the email to be sent :S
-                    $content->seotitle = $content->gen_seotitle($this->request->post('title'));
 
                 try 
                 {
@@ -206,16 +204,23 @@ class Controller_Panel_Content extends Auth_Controller {
         $content = new Model_Content($id);
         
         if ($content->loaded())
-        {
-            try
+        {   
+            //deleting default locale emails is not allowed
+            if ($content->type == 'email' AND $content->locale == i18n::$locale_default)
             {
-                $content->delete();
-                $this->template->content = 'OK';
+                Alert::set(Alert::INFO, sprintf(__('Sorry, deleting %s locale emails is not allowed'), i18n::$locale_default));
+                HTTP::redirect(Route::url('oc-panel',array('controller'  => 'content','action'=>'list')).'?type='.$content->type.'&locale_select='.$content->locale);
             }
-            catch (Exception $e)
-            {
-                 Alert::set(Alert::ERROR, $e->getMessage());
-            }
+            else
+                try
+                {
+                    $content->delete();
+                    $this->template->content = 'OK';
+                }
+                catch (Exception $e)
+                {
+                     Alert::set(Alert::ERROR, $e->getMessage());
+                }
         }
        
     }
